Introduce the Partial type, reorganize What4.Partial
