+++ crucible-llvm/src/Lang/Crucible/LLVM/MemModel/Partial.hs
@@ -74,8 +74,10 @@ import           Lang.Crucible.Panic (panic)
 import           What4.Interface (Pred, IsExprBuilder)
 import qualified What4.Interface as W4I
 import qualified What4.InterpretedFloatingPoint as W4IFP
-import           What4.Partial (PartExpr(..), AssertionTree(..))
-import qualified What4.Partial as W4P
+import           What4.Partial.PartExpr (PartExpr(..))
+import qualified What4.Partial.PartExpr as W4P hiding (addCondition)
+import           What4.Partial.AssertionTree (AssertionTree(..))
+import           What4.Partial.AssertionTree as W4P
 
 ------------------------------------------------------------------------
 -- ** PartLLVMVal
