+++ crucible/src/Lang/Crucible/Simulator/RegValue.hs
@@ -65,7 +65,7 @@ import qualified Data.Parameterized.Context as Ctx
 import           What4.FunctionName
 import           What4.Interface
 import           What4.InterpretedFloatingPoint
-import           What4.Partial
+import           What4.Partial.PartExpr
 import           What4.WordMap
 
 import           Lang.Crucible.FunctionHandle
