+++ what4/what4.cabal
@@ -63,6 +63,8 @@ library
     What4.Interface
     What4.InterpretedFloatingPoint
     What4.Partial
+    What4.Partial.AssertionTree
+    What4.Partial.PartExpr
     What4.ProblemFeatures
     What4.ProgramLoc
     What4.SatResult
